+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # TLSPROXY Release Notes
 
+## next
+
+* Add support for the PROXY protocol defined by HAProxy. https://github.com/haproxy/haproxy/blob/master/doc/proxy-protocol.txt
+
 ## v0.2.0
 
 * Add support for the HTTP/2 protocol in reverse proxy mode. This only works when the backend actually supports HTTP/2. So, it is not enabled by default. To enable it, set `alpnProtos: [h2, http/1.1]` on the backend explicitly.